+++ CHANGELOG.md
@@ -16,6 +16,10 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - C#: parse __makeref, __reftype, __refvalue (#3364)
 - Java: parsing of dots inside function annotations with brackets (#3389)
 - Do not pretend that short-circuit Boolean AND and OR operators are commutative (#3399)
+- metavariable-pattern: Fix crash when nesting a non-generic pattern within
+  a generic rule
+- metavariable-pattern: Fix parse info when matching content of a metavariable
+  under a different language
 
 ### Changed